add ci check for git deploy plugin (unrelated)
